requirements.txt
fastapi
pydantic
uvicorn
sentence-transformers
transformers
pinecone
dotenv
hf_xet
sqlmodel
psycopg2
py-grpc-prometheus
python-logging-loki
psutil

opentelemetry-instrumentation-fastapi
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc